Add a new test deploy with randomness.
The randomness this deploy introduces helps to validate that when preparing
operations in parallel the order of the operations are called on each host
has no impact on the result.
