+++ tests/test_cli/test_cli.py
@@ -1,40 +1,28 @@
 from os import path
-from random import shuffle
 from unittest import TestCase
 
-from click.testing import CliRunner
-
-import pyinfra
-
 from pyinfra import pseudo_state
-from pyinfra_cli.main import _main, cli
+from pyinfra_cli.main import _main
 
+from .util import run_cli
 from ..paramiko_util import PatchSSHTestCase
-
-
-def run_cli(*arguments):
-    pyinfra.is_cli = True
-    runner = CliRunner(mix_stderr=False)
-    result = runner.invoke(cli, arguments)
-    pyinfra.is_cli = False
-    return result
 
 
 class TestCliEagerFlags(TestCase):
     def test_print_help(self):
         result = run_cli('--version')
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
         result = run_cli('--help')
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_print_facts_list(self):
         result = run_cli('--facts')
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_print_operations_list(self):
         result = run_cli('--operations')
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
 
 class TestDeployCli(PatchSSHTestCase):
@@ -46,8 +34,8 @@ class TestDeployCli(PatchSSHTestCase):
             '@local',
             'not-a-file.py',
         )
-        assert result.exit_code == 1, result.stderr
-        assert 'No deploy file: not-a-file.py' in result.stderr
+        assert result.exit_code == 1, result.stdout
+        assert 'No deploy file: not-a-file.py' in result.stdout
 
 
 class TestOperationCli(PatchSSHTestCase):
@@ -56,7 +44,7 @@ class TestOperationCli(PatchSSHTestCase):
             path.join('tests', 'deploy', 'inventories', 'inventory.py'),
             'not_a_module.shell',
         )
-        assert result.exit_code == 1, result.stderr
+        assert result.exit_code == 1, result.stdout
         assert 'No such module: not_a_module'
 
     def test_invalid_operation_function(self):
@@ -64,7 +52,7 @@ class TestOperationCli(PatchSSHTestCase):
             path.join('tests', 'deploy', 'inventories', 'inventory.py'),
             'server.not_an_operation',
         )
-        assert result.exit_code == 1, result.stderr
+        assert result.exit_code == 1, result.stdout
         assert 'No such operation: server.not_an_operation'
 
     def test_deploy_inventory(self):
@@ -73,7 +61,7 @@ class TestOperationCli(PatchSSHTestCase):
             'server.shell',
             '--debug-data',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_deploy_operation(self):
         result = run_cli(
@@ -81,7 +69,7 @@ class TestOperationCli(PatchSSHTestCase):
             'server.shell',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_deploy_operation_with_all(self):
         result = run_cli(
@@ -89,7 +77,7 @@ class TestOperationCli(PatchSSHTestCase):
             'server.shell',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_deploy_operation_json_args(self):
         result = run_cli(
@@ -97,7 +85,7 @@ class TestOperationCli(PatchSSHTestCase):
             'server.shell',
             '[["echo hi"], {}]',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
 
 class TestFactCli(PatchSSHTestCase):
@@ -107,8 +95,8 @@ class TestFactCli(PatchSSHTestCase):
             'fact',
             'server.Os',
         )
-        assert result.exit_code == 0, result.stderr
-        assert '"somehost": null' in result.stderr
+        assert result.exit_code == 0, result.stdout
+        assert '"somehost": null' in result.stdout
 
     def test_get_fact_with_kwargs(self):
         result = run_cli(
@@ -117,8 +105,8 @@ class TestFactCli(PatchSSHTestCase):
             'files.File',
             'path=.',
         )
-        assert result.exit_code == 0, result.stderr
-        assert '"somehost": null' in result.stderr
+        assert result.exit_code == 0, result.stdout
+        assert '"somehost": null' in result.stdout
 
     def test_invalid_fact_module(self):
         result = run_cli(
@@ -126,8 +114,8 @@ class TestFactCli(PatchSSHTestCase):
             'fact',
             'not_a_module.NotAFact',
         )
-        assert result.exit_code == 1, result.stderr
-        assert 'No such module: not_a_module' in result.stderr
+        assert result.exit_code == 1, result.stdout
+        assert 'No such module: not_a_module' in result.stdout
 
     def test_invalid_fact_class(self):
         result = run_cli(
@@ -135,8 +123,8 @@ class TestFactCli(PatchSSHTestCase):
             'fact',
             'server.NotAFact',
         )
-        assert result.exit_code == 1, result.stderr
-        assert 'No such fact: server.NotAFact' in result.stderr
+        assert result.exit_code == 1, result.stdout
+        assert 'No such fact: server.NotAFact' in result.stdout
 
     def test_get_facts_legacy(self):
         result = run_cli(
@@ -144,8 +132,8 @@ class TestFactCli(PatchSSHTestCase):
             'fact',
             'os',
         )
-        assert result.exit_code == 0, result.stderr
-        assert '"somehost": null' in result.stderr
+        assert result.exit_code == 0, result.stdout
+        assert '"somehost": null' in result.stdout
 
 
 class TestExecCli(PatchSSHTestCase):
@@ -156,7 +144,7 @@ class TestExecCli(PatchSSHTestCase):
             '--',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_exec_command_with_options(self):
         result = run_cli(
@@ -170,7 +158,7 @@ class TestExecCli(PatchSSHTestCase):
             '--',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_exec_command_with_serial(self):
         result = run_cli(
@@ -180,7 +168,7 @@ class TestExecCli(PatchSSHTestCase):
             '--',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_exec_command_with_no_wait(self):
         result = run_cli(
@@ -190,7 +178,7 @@ class TestExecCli(PatchSSHTestCase):
             '--',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_exec_command_with_debug_operations(self):
         result = run_cli(
@@ -200,7 +188,7 @@ class TestExecCli(PatchSSHTestCase):
             '--',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_exec_command_with_debug_facts(self):
         result = run_cli(
@@ -210,7 +198,7 @@ class TestExecCli(PatchSSHTestCase):
             '--',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
+        assert result.exit_code == 0, result.stdout
 
     def test_exec_command_with_debug_data_limit(self):
         result = run_cli(
@@ -221,97 +209,7 @@ class TestExecCli(PatchSSHTestCase):
             '--',
             'echo hi',
         )
-        assert result.exit_code == 0, result.stderr
-
-
-class TestCliDeployState(PatchSSHTestCase):
-    def _assert_op_data(self, correct_op_name_and_host_names):
-        state = pseudo_state
-        op_order = state.get_op_order()
-
-        assert (
-            len(correct_op_name_and_host_names) == len(op_order)
-        ), 'Incorrect number of operations detected'
-
-        for i, (correct_op_name, correct_host_names) in enumerate(
-            correct_op_name_and_host_names,
-        ):
-            op_hash = op_order[i]
-            op_meta = state.op_meta[op_hash]
-
-            assert list(op_meta['names'])[0] == correct_op_name
-
-            for host in state.inventory:
-                op_hashes = state.meta[host]['op_hashes']
-                if correct_host_names is True or host.name in correct_host_names:
-                    self.assertIn(op_hash, op_hashes)
-                else:
-                    self.assertNotIn(op_hash, op_hashes)
-
-    def test_deploy(self):
-        task_file_path = path.join('tests', 'deploy', 'tasks', 'a_task.py')
-        nested_task_path = path.join('tests', 'deploy', 'tasks', 'another_task.py')
-        correct_op_name_and_host_names = [
-            ('First main operation', True),  # true for all hosts
-            ('Second main operation', ('somehost',)),
-            ('{0} | First task operation'.format(task_file_path), ('anotherhost',)),
-            ('{0} | Task order loop 1'.format(task_file_path), ('anotherhost',)),
-            ('{0} | 2nd Task order loop 1'.format(task_file_path), ('anotherhost',)),
-            ('{0} | Task order loop 2'.format(task_file_path), ('anotherhost',)),
-            ('{0} | 2nd Task order loop 2'.format(task_file_path), ('anotherhost',)),
-            (
-                '{0} | {1} | Second task operation'.format(task_file_path, nested_task_path),
-                ('anotherhost',),
-            ),
-            ('{0} | First task operation'.format(task_file_path), True),
-            ('{0} | Task order loop 1'.format(task_file_path), True),
-            ('{0} | 2nd Task order loop 1'.format(task_file_path), True),
-            ('{0} | Task order loop 2'.format(task_file_path), True),
-            ('{0} | 2nd Task order loop 2'.format(task_file_path), True),
-            ('{0} | {1} | Second task operation'.format(task_file_path, nested_task_path), True),
-            ('My deploy | First deploy operation', True),
-            ('My deploy | My nested deploy | First nested deploy operation', True),
-            ('My deploy | Second deploy operation', True),
-            ('Loop-0 main operation', True),
-            ('Loop-1 main operation', True),
-            ('Third main operation', True),
-            ('Order loop 1', True),
-            ('2nd Order loop 1', True),
-            ('Order loop 2', True),
-            ('2nd Order loop 2', True),
-            ('Final limited operation', ('somehost',)),
-        ]
-
-        # Run 3 iterations of the test - each time shuffling the order of the
-        # hosts - ensuring that the ordering has no effect on the operation order.
-        for _ in range(3):
-            pseudo_state.reset()
-
-            hosts = ['somehost', 'anotherhost', 'someotherhost']
-            shuffle(hosts)
-
-            result = run_cli(
-                ','.join(hosts),
-                path.join('tests', 'deploy', 'deploy.py'),
-            )
-            assert result.exit_code == 0, result.stderr
-
-            self._assert_op_data(correct_op_name_and_host_names)
-
-    def test_interdependent_deploy(self):
-        pseudo_state.reset()
-
-        result = run_cli(
-            'somehost',
-            path.join('tests', 'deploy', 'deploy_interdependent.py'),
-        )
-        assert result.exit_code == 0, result.stderr
-
-        # Check every operation had commands/changes - this ensures that each
-        # combo (add/remove/add) always had changes.
-        for host, ops in pseudo_state.ops.items():
-            for _, op in ops.items():
-                assert len(op['commands']) > 0
+        assert result.exit_code == 0, result.stdout
 
 
 class TestDirectMainExecution(PatchSSHTestCase):
